+++ src/libretro/drivers/video/software/__init__.py
@@ -1,7 +1,2 @@
 from .array import *
 from .base import *
-
-try:
-    from .pillow import *
-except ImportError:
-    pass